MCKIN-21527 Problem Builder code review changes
