+++ problem_builder/public/js/mentoring.js
@@ -3,7 +3,6 @@ function MentoringBlock(runtime, element) {
     var gettext = window.ProblemBuilderXBlockI18N.gettext;
     var ngettext = window.ProblemBuilderXBlockI18N.ngettext;
 
-    var attemptsTemplate = _.template($('#xblock-attempts-template').html());
     var data = $('.mentoring', element).data();
     var children = runtime.children(element);
     var step = data.step;
@@ -67,13 +66,12 @@ function MentoringBlock(runtime, element) {
 
     function renderAttempts() {
         var data = $('.attempts', element).data();
-        if (data != undefined && data.max_attempts > 0) {
+        if (data != undefined && _.isNumber(data.max_attempts) && data.max_attempts > 0) {
            var message = _.template(
               ngettext("You have used {num_used} of 1 submission.", "You have used {num_used} of {max_attempts} submissions.", data.max_attempts),
               {num_used: _.min([data.num_attempts, data.max_attempts]), max_attempts: data.max_attempts}, {interpolate: /\{(.+?)\}/g}
             );
-           data.message = message;
-           $('.attempts', element).html(attemptsTemplate(data));
+           $('.attempts', element).html("<span>" + message + "</span>");
         }
     }
