--- problem_builder/templates/html/mentoring_attempts.underscore
@@ -1,7 +1,0 @@
-<script type="text/template" id="xblock-attempts-template">
-  <% if (_.isNumber(max_attempts) && max_attempts > 0) {{ %>
-  <span>
-  	<%= message  %>
-  </span>
-  <% }} %>
-</script>
