+++ setup.py
@@ -28,7 +28,7 @@ from setuptools.command.install import install
 
 # Constants #########################################################
 
-VERSION = '3.4.23'
+VERSION = '3.5.0'
 
 # Functions #########################################################
